rft: Wpf重构 拆分模拟器控制相关
